requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.30.1
deepgram-sdk~=3.4.0
google-genai~=0.1.0
python-dotenv~=1.0.1
aiofiles~=23.2.1
pydantic~=2.7.4
sqlalchemy~=2.0.23
aiosqlite~=0.19.0
requests~=2.31.0
